Initial support for screen ambiance to allow websocket connections. Blazor WASM projects to consume and display data from screen ambiance websocket. Replace Iot.Device.Bindings with a local minimal version because it makes asp.net not work?? This also allows me to finally be rid of System.Drawing and only use ImageSharp in Iot.Device.Bindings.

+++ HueScreenAmbience/DXGICaptureScreen/DxEnumerate.cs
@@ -16,7 +16,7 @@ namespace HueScreenAmbience.DXGICaptureScreen
 		public string Name { get; set; }
 		public int Width { get; set; }
 		public int Height { get; set; }
-		public int RefreshRate { get; set; }
+		public double RefreshRate { get; set; }
 		public string Format { get; set; }
 		public int Bpp { get; set; }
 		public string ColorSpace { get; set; }
